+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+ - Remove the requirement that `ML99_variadicsForEach(I)` can accept at most 63 arguments.
+
 ## [1.10.0] - 2021-09-14
 
 ### Added